Move insight.py to pipeline.py
