+++ afqinsight/__init__.py
@@ -2,6 +2,6 @@
 from . import datasets  # noqa
 from . import utils  # noqa
 from .datasets import *  # noqa
-from .insight import *  # noqa
+from .pipeline import *  # noqa
 from .transform import *  # noqa
 from ._version import version as __version__  # noqa
